+++ Sloth/Common/TitleInput/TitleActionSheet.swift
@@ -1,0 +1,18 @@
+//
+//  TitleActionSheet.swift
+//  Sloth
+//
+//  Created by 심지원 on 2022/05/15.
+//
+
+import UIKit
+
+class TitleActionSheet: TitleInput {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
